関数宣言を整理した
